refactor: move internal/checkmetrics to internal/check/metrics
