+++ internal/check/check.go
@@ -17,7 +17,7 @@ import (
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 	authzGraph "github.com/openfga/language/pkg/go/graph"
 
-	"github.com/openfga/openfga/internal/checkmetrics"
+	"github.com/openfga/openfga/internal/check/metrics"
 	"github.com/openfga/openfga/internal/concurrency"
 	"github.com/openfga/openfga/internal/iterator"
 	"github.com/openfga/openfga/internal/modelgraph"
@@ -159,7 +159,7 @@ func (r *Resolver) isCached(consistency openfgav1.ConsistencyPreference, key key
 	if consistency == openfgav1.ConsistencyPreference_HIGHER_CONSISTENCY {
 		return nil, false
 	}
-	checkmetrics.CacheLookupCounter.Inc()
+	metrics.CacheLookupCounter.Inc()
 	v := r.cache.Get(key)
 	if v == nil {
 		return nil, false
@@ -169,10 +169,10 @@ func (r *Resolver) isCached(consistency openfgav1.ConsistencyPreference, key key
 		return nil, false
 	}
 	if !res.LastModified.After(r.lastCacheInvalidationTime) {
-		checkmetrics.CacheInvalidHitCounter.Inc()
+		metrics.CacheInvalidHitCounter.Inc()
 		return nil, false
 	}
-	checkmetrics.CacheHitCounter.Inc()
+	metrics.CacheHitCounter.Inc()
 	return res.Res, true
 }
